[Core] Move member-shared view model bits to new base
