+++ Xamarin.PropertyEditing/Xamarin.PropertyEditing.csproj
@@ -70,6 +70,7 @@
     <Compile Include="ValueSource.cs" />
     <Compile Include="ViewModels\ArrangeModeViewModel.cs" />
     <Compile Include="ViewModels\ConstrainedPropertyViewModel.cs" />
+    <Compile Include="ViewModels\EditorViewModel.cs" />
     <Compile Include="ViewModels\EnumPropertyViewModel.cs" />
     <Compile Include="Extensions.cs" />
     <Compile Include="ViewModels\NumericPropertyViewModel.cs" />
